vnc: fix tests
Minor issues.
