+++ src/vnc/protocol_test.go
@@ -6,7 +6,6 @@ package vnc
 
 import (
 	"bytes"
-	"fmt"
 	"reflect"
 	"testing"
 )
@@ -32,8 +31,8 @@ func TestWriteRead(t *testing.T) {
 		},
 		&FramebufferUpdateRequest{
 			Incremental: 1,
-			XPosition:   2,
-			YPosition:   3,
+			X:           2,
+			Y:           3,
 			Width:       4,
 			Height:      5,
 		},
@@ -83,54 +82,5 @@ func TestReadBuffer(t *testing.T) {
 			t.Fatalf("read message failed -- %v", err)
 		}
 		t.Logf("%#v\n", msg)
-	}
-}
-
-func TestStringParse(t *testing.T) {
-	vals := []fmt.Stringer{
-		&KeyEvent{
-			DownFlag: 0, Key: 0xff08,
-		},
-		&KeyEvent{
-			DownFlag: 1, Key: 0x0030,
-		},
-		&KeyEvent{
-			DownFlag: 1, Key: 0x0041,
-		},
-		&KeyEvent{
-			DownFlag: 1, Key: 0xffe1,
-		},
-		&PointerEvent{
-			ButtonMask: 0, XPosition: 100, YPosition: 200,
-		},
-		&PointerEvent{
-			ButtonMask: 0, XPosition: 105, YPosition: 200,
-		},
-		&PointerEvent{
-			ButtonMask: 1, XPosition: 105, YPosition: 205,
-		},
-	}
-
-	for _, want := range vals {
-		var got interface{}
-		var err error
-
-		s := want.String()
-
-		switch want.(type) {
-		case *KeyEvent:
-			got, err = parseKeyEvent(s)
-		case *PointerEvent:
-			got, err = parsePointerEvent(s)
-		}
-
-		if err != nil {
-			t.Errorf("parse failed -- %v", err)
-			continue
-		}
-
-		if !reflect.DeepEqual(got, want) {
-			t.Errorf("struct aren't equal -- got: %v, want: %v", got, want)
-		}
 	}
 }
